Rename: 시간기록함수 이동
시간기록 함수를 댓글 컴포넌트로 이동했습니다.
