--- src/components/timeForToday.jsx
@@ -1,27 +1,0 @@
-export const timeForToday = (createTime) => {
-    const today = new Date();
-    const timeValue = new Date(createTime);
-  
-    const betweenTime = Math.floor(
-      (today.getTime() - timeValue.getTime()) / 1000 / 60,
-    );
-
-    if (betweenTime < 1) return '방금 전';
-    if (betweenTime < 60) {
-      return `${betweenTime}분 전`;
-    }
-  
-    const betweenTimeHour = Math.floor(betweenTime / 60);
-
-    if (betweenTimeHour < 24) {
-      return `${betweenTimeHour}시간 전`;
-    }
-  
-    const betweenTimeDay = Math.floor(betweenTime / 60 / 24);
-    
-    if (betweenTimeDay < 365) {
-      return `${betweenTimeDay}일 전`;
-    }
-  
-    return `${Math.floor(betweenTimeDay / 365)}년 전`;
-  };
